Start to rework the config system with a config file
